requirements.txt
fastapi==0.129.0
uvicorn==0.38.0
pydantic==2.12.5
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib==1.7.4

redis==6.4.0
rq==2.6.1

torch==2.10.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
yt-dlp==2026.2.21
ollama==0.6.1
